requirements.txt
streamlit
requests
pandas
fastapi
uvicorn
qdrant-client
google-genai
pydantic-settings
python-dotenv
